Static library for iOS
Modified the macosx Xcode project to include an iOS target. This is just much easier to include as a sub project instead of building separate libraries and lipoing them together. Used the config.h file from an arm64 build, and it seems to work fine for arm32 builds. One point of discrepancy is the SIZEOF_LONG etc defined in the config file as 8 which is incorrect for arm7, but they don't seem to be used anywhere. Good thing.
Changed the <config.h> to "config.h" in all places. Removed the '\param void' doc comments. Too many other warnings to worry about for now. Wow this is just layers of old nonsense. Will see about a full pass cleanup later